+++ HorseRacingTournamentManagementSystem_0/Modules/Users/DTOs/CreateUserRequestDto.cs
@@ -19,5 +19,5 @@ public class CreateUserRequestDto
     // Role-specific fields
     public int? ExperienceYear { get; set; }
     public int? ExpYears { get; set; }
-    public int? TotalPoints { get; set; }
+    public double? TotalPoints { get; set; }
 }
